+++ transport/src/main/java/com/cooperative/transport/repositories/StatutReservationRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.cooperative.transport.entities.StatutReservation;
 
-public interface StatutReservationRepository extends JpaRepository<StatutReservation, Long> {
+public interface StatutReservationRepository extends JpaRepository<StatutReservation, Integer> {
 
     public Optional<StatutReservation> findByLibelle(String libelle);
 }
